requirements.txt
selenium>=4.16.0
python-dotenv>=1.0.0
configparser>=6.0.0
mysql-connector-python>=8.0.0 